requirements.txt
flask==3.1.0
requests==2.32.3
python-whois==0.9.4
beautifulsoup4==4.13.3
gunicorn==23.0.0
